+++ src/private/shared/GetNovaResolvedProjectPackageSettings.ps1
@@ -7,20 +7,9 @@ function Get-NovaResolvedProjectPackageSettings {
         [Parameter(Mandatory)][string]$ProjectRoot
     )
 
-    $packageSettings = if ($ProjectData.ContainsKey('Package') -and $ProjectData['Package'] -is [hashtable]) {
-        [ordered]@{} + $ProjectData['Package']
-    }
-    else {
-        [ordered]@{}
-    }
-
-    Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Enabled' -Value $true
+    $packageSettings = Get-NovaProjectPackageSettingsTable -ProjectData $ProjectData
+    $packageSettings['OutputDirectory'] = Get-NovaResolvedProjectPackageOutputDirectorySettings -PackageSettings $packageSettings -ProjectRoot $ProjectRoot
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Id' -Value $ProjectData['ProjectName'] -TreatWhitespaceAsMissing
-    Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'OutputDirectory' -Value 'artifacts/packages' -TreatWhitespaceAsMissing
-
-    if (-not [System.IO.Path]::IsPathRooted("$( $packageSettings['OutputDirectory'] )")) {
-        $null = $packageSettings['OutputDirectory'] = [System.IO.Path]::Join($ProjectRoot, "$( $packageSettings['OutputDirectory'] )")
-    }
 
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'PackageFileName' -Value "$( $packageSettings['Id'] ).$( $ProjectData['Version'] ).nupkg" -TreatWhitespaceAsMissing
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Authors' -Value $ManifestSettings['Author']